requirements.txt
fastapi
pyodbc
python-decouple
